src/requirements.txt
requests==2.21.0
Flask==2.0.3
oauthlib==3.0.1
pyOpenSSL==23.0.0
Flask-Login==0.4.1
Werkzeug==2.0.3
MarkupSafe==2.1.1
Jinja2==3.1.1
google-api-python-client==1.10.0
google-auth-httplib2
google-auth-oauthlib
polling2